+++ examples/custom_ui_widgets.rs
@@ -2,6 +2,8 @@
 //! This example shows how to use the add_ui_callback method to create
 //! custom windows, panels, and other UI elements using egui.
 //! This example requires the `viewer-ui` cargo feature to be enabled.
+//! 
+//! This example uses the viewer in single-threaded fashion.
 
 use std::time::Duration;
 use std::cell::RefCell;