+++ docs/guide/source/changelog.rst
@@ -24,15 +24,15 @@ update of MuJoCo alone can increase the major version.
   - New feature added for compatibility purposes: ``renderer-winit-fallback``.
     The feature is enabled by default.
 
-- Info menu (F2) displaying smoothed FPS, simulation time, and used memory.
-- Realtime warning (F4) for showing the realtime factor [%] when the simulation's sync time deviates at least 2 %
-  from the model's configured timestep.
 - Changes to :ref:`mj_rust_viewer`:
 
   - Added :docs-rs:`~~mujoco_rs::viewer::<struct>MjViewer::<method>add_ui_callback` for custom UI widgets.
     This allows users to create custom windows, panels, and other UI elements using `egui <https://github.com/emilk/egui>`_.
     See :ref:`custom_ui_widgets` for more information.
 
+  - Info menu (F2) displaying smoothed FPS, simulation time, and used memory.
+  - Realtime warning (F4) for showing the realtime factor [%] when the simulation's sync time deviates at least 2 %
+    from the model's configured timestep.
 
   - Separate syncing and rendering logic:
 
@@ -63,6 +63,14 @@ update of MuJoCo alone can increase the major version.
 - New examples:
 
   - :gh-example:`rust_viewer_threaded.rs`.
+
+- Deprecated:
+
+  - :docs-rs:`~~mujoco_rs::viewer::<struct>MjViewer::<method>user_scene` and
+    :docs-rs:`~~mujoco_rs::viewer::<struct>MjViewer::<method>user_scene_mut`.
+    They are replaced with
+    :docs-rs:`~~mujoco_rs::viewer::<struct>ViewerSharedState::<method>user_scene` and
+    :docs-rs:`~~mujoco_rs::viewer::<struct>ViewerSharedState::<method>user_scene_mut`.